Fixed cutout issues
